+++ src/evaluators/response/PHAL_ResponseSquaredL2Difference_Def.hpp
@@ -100,7 +100,7 @@ void PHAL::ResponseSquaredL2DifferenceBase<EvalT, Traits, SourceScalarT, TargetS
   // Zero out local response
   PHAL::set(this->local_response_eval, 0.0);
 
-  for (int cell=0; cell<workset.numCells; ++cell)
+  for (size_t cell=0; cell<workset.numCells; ++cell)
   {
     ScalarT sum = 0;
     for (int qp=0; qp<numQPs; ++qp)
@@ -113,12 +113,12 @@ void PHAL::ResponseSquaredL2DifferenceBase<EvalT, Traits, SourceScalarT, TargetS
           sq += std::pow(sourceField(cell,qp)-(target_value ? target_value_val : targetField(cell,qp)),2);
           break;
         case 1:
-          for (int j=0; j<dims[2]; ++j)
+          for (size_t j=0; j<dims[2]; ++j)
             sq += std::pow(sourceField(cell,qp,j)-(target_value ? target_value_val : targetField(cell,qp,j)),2);
           break;
         case 2:
-          for (int j=0; j<dims[2]; ++j)
-            for (int k=0; k<dims[3]; ++k)
+          for (size_t j=0; j<dims[2]; ++j)
+            for (size_t k=0; k<dims[3]; ++k)
               sq += std::pow(sourceField(cell,qp,j,k)-(target_value ? target_value_val : targetField(cell,qp,j,k)),2);
           break;
       }
